backend/posts_service/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
confluent-kafka==2.1.1
python-dotenv==1.0.0
django-environ==0.10.0